refactor(master): delegate periodic ops to backends
Add IFilesystemOperations hooks for the remaining periodic work:
fsTestPeriodicTick, fsTestBackgroundStep, fsTestGetData,
fsTestGetDefectiveNodes, fsTestOnNodeRemoved, and
backgroundChecksumStep. Add FsTestReport as the internal carrier
for the existing fs_test_getdata() fields.

Move the in-memory file-test scanner state, scan loop, defective
node map, report rendering, and remove-node cleanup from
filesystem_periodic.cc into FilesystemOperationsBase. The Master
implementation keeps the existing scan order, timing, log lines,
counter math, report text, and gDefectiveNodes semantics.

Make filesystem_periodic.cc scheduler glue only. Keep the public
C-style entry points used by matoclserv.cc and filesystem_node.cc,
but route them through gFSOperations. Update fsnodes_periodic_remove
to take a FilesystemOperationContext so KV backends can clear
backend file-test state in the same transaction as node deletion.

Move the per-loop metadata checksum recalculation body behind
backgroundChecksumStep, matching the existing start hook. Keep the
Base implementation unchanged and let KV backends override it.

Add metadataChecksumSupported() and readBackendPeriodicConfig() to
the operations interface. Gate ChecksumUpdater changelog emission
on metadataChecksumSupported() so backends without Master-Shadow
metadata checksums do not emit meaningless CHECKSUM(...):0 entries.

Add the shared FILETEST_* key constants in kv_common_keys.h:
FILETEST_LEASE, FILETEST_CURSOR, FILETEST_ACTIVE_STATS,
FILETEST_PUBLISHED_STATS, FILETEST_PUBLISHED_REPORT,
FILETEST_DEFECTIVE_<inode>, and
FILETEST_REPORT_<scan_generation><inode>. This commit only adds the
master-side seam and shared key names; it does not implement FDB
scanner behavior in this repo.
