+++ src/master/README.md
@@ -273,14 +273,32 @@ Jobs support cancellation and completion callbacks.
 Several maintenance routines run on timer-driven or per-loop schedules,
 registered in `fs_periodic_master_init()` (see `filesystem_periodic.cc`):
 
-| Operation | Schedule | Purpose |
-|-----------|----------|---------|
-| **File integrity test** (`fs_periodic_file_test` / `fs_background_file_test`) | Every second (timer) + every loop (background) | Scans the entire node hash table in a configurable cycle time (`FILE_TEST_LOOP_MIN_TIME`, default 3600s). For each file node, checks chunk availability and copy counts. For each directory, validates parent-child pointer consistency. Builds a `gDefectiveNodes` map of inodes with unavailable chunks, under-goal chunks, or structural errors. |
-| **Background task processing** (`fs_background_task_manager_work`) | Every loop | Drives the `TaskManager`, processing a batch of tasks (snapshots, recursive removes, goal/trashtime changes) per iteration. |
-| **Checksum recalculation** (`fs_background_checksum_recalculation_a_bit`) | Every loop | Incrementally recalculates metadata checksums (nodes, xattrs, chunks) in the background, progressing through steps at a speed limit per iteration. |
-| **Trash cleanup** (`fs_periodic_emptytrash`) | Every 100ms | Purges expired trash entries whose deletion timestamp has passed. |
-| **Reserved file cleanup** (`fs_periodic_emptyreserved`) | Configurable period in ms (`EMPTY_RESERVED_FILES_PERIOD_MSECONDS`); `0` disables | Force-releases reserved files (deleted-but-still-open files) from all owning sessions, even if sessions are still active; enabling it can disrupt clients that still hold those files. |
-| **Chunk maintenance** (in `chunks.cc`) | Periodic | Handles chunk replication, deletion of excess copies, and rebalancing across chunkservers. |
+- **File integrity test** (`fs_periodic_file_test` / `fs_background_file_test`):
+  runs every second (timer) and every loop (background). Scans the entire node
+  hash table in a configurable cycle time (`FILE_TEST_LOOP_MIN_TIME`, default
+  3600s). For each file node, checks chunk availability and copy counts. For
+  each directory, validates parent-child pointer consistency. Builds a
+  `gDefectiveNodes` map of inodes with unavailable chunks, under-goal chunks,
+  or structural errors.
+- **Background task processing** (`fs_background_task_manager_work`):
+  runs every loop. Drives the `TaskManager`, processing a batch of tasks
+  (snapshots, recursive removes, goal/trashtime changes) per iteration.
+- **Checksum recalculation** (`fs_background_checksum_recalculation_a_bit`):
+  runs every loop. Incrementally recalculates metadata checksums (nodes,
+  xattrs, chunks) in the background, progressing through steps at a speed
+  limit per iteration.
+- **Trash cleanup** (`fs_periodic_emptytrash`):
+  runs every 100ms. Purges expired trash entries whose deletion timestamp has
+  passed.
+- **Reserved file cleanup** (`fs_periodic_emptyreserved`):
+  runs on a configurable period in ms
+  (`EMPTY_RESERVED_FILES_PERIOD_MSECONDS`), where `0` disables it.
+  Force-releases reserved files (deleted-but-still-open files) from all owning
+  sessions, even if sessions are still active; enabling it can disrupt clients
+  that still hold those files.
+- **Chunk maintenance** (in `chunks.cc`):
+  runs periodically. Handles chunk replication, deletion of excess copies, and
+  rebalancing across chunkservers.
 
 ## Initialization Sequence
